Breakout DrawableBase
